working flask app

+++ gene_api.py
@@ -60,10 +60,8 @@ def handle_data():
     """
     if request.method == 'GET':
         output_list = []
-        try:
-            len(rd.keys()) >= 1
-        except Exception:
-            return ("No data in the database to retrieve\n")
+        if len(rd.keys()) < 1:
+            return ("No data in the database to retrieve. Please use a POST route first.\n")
         for item in rd.keys():
             output_list.append(json.loads(rd.get(item)))
         return output_list
@@ -72,16 +70,14 @@ def handle_data():
         for item in data:
             key = f'{item["hgnc_id"]}'
             rd.set(key, json.dumps(item))
-        return f'Data loaded\n'
+        return f'Data loaded.\n'
     elif request.method == 'DELETE':
-        try:
-            len(rd.keys()) >= 1
-        except Exception:
-            return ("No data in the database to delete\n")
+        if len(rd.keys()) < 1:
+            return ("No data in the database to delete.\n")
         rd.flushdb()
-        return f'Data deleted, there are {len(rd.keys())} keys in the db\n'
+        return f'Data deleted, there are {len(rd.keys())} keys in the db.\n'
     else:
-        return 'The method you tried does not work\n'
+        return 'The method you tried does not work.\n'
 
 @app.route('/genes/<string:hgnc_id>', methods = ['GET'])
 def get_gene(hgnc_id: str) -> dict:
@@ -94,10 +90,8 @@ def get_gene(hgnc_id: str) -> dict:
     Returns:
         items (dict): Dictionary with all data for the hgnc_id.
     """
-    try:
-        len(rd.keys()) >= 1
-    except Exception:
-        return ("No data in the database\n")
+    if len(rd.keys()) < 1:
+        return ("No data in the database. Please use a POST route first.\n")
     items = json.loads(rd.get(hgnc_id))
     return items
 
@@ -113,10 +107,8 @@ def get_genes() -> list:
         output (list): List of all hgnc_ids.
     """
     output = []
-    try:
-        len(rd.keys()) >= 1
-    except Exception:
-        return ("No data available in the database\n")
+    if len(rd.keys()) < 1:
+        return ("No data available in the database. Please use a POST route first.\n")
     return rd.keys()
 
 @app.route('/image', methods = ['POST','GET', 'DELETE'])
@@ -132,10 +124,8 @@ def get_image():
         (file_bytes): Returns the file_bytes of the image in the POST route.
         send_file(): Returns the image to the user in the GET route.
     """
-    try:
-        len(rd.keys()) >= 1
-    except Exception:
-        return ("No data in the database\n")
+    if len(rd.keys()) < 1:
+        return ("No data in the database. Please use a POST route first.\n")
     years = []
     counts = []
     if request.method == 'POST':
@@ -144,35 +134,37 @@ def get_image():
             year = gene['date_approved_reserved'][0:4]
             years.append(year)
         yeard = dict(Counter(years))
-        years = list(yeard.keys())
-        counts = list(yeard.values())
+        yeard = dict(sorted(yeard.items()))
+        y = []
+        c = []
+        for item in yeard:
+            y.append(item)
+            c.append(yeard[item])
         plt.figure(figsize=(28,6))
-        plt.bar(years, counts, width = 0.35)
+        plt.bar(y, c, width = 0.35)
         plt.xlabel("Years")
         plt.ylabel("Number of Entries Approved")
         plt.title("Genes Approved Each Year")
         plt.savefig('approvalyears.png')
         file_bytes = open('./approvalyears.png', 'rb').read()
         rd1.set('genes_approved', file_bytes)
-        return ("Image created\n")
+        return ("Image created.\n")
     elif request.method == 'GET':
         path = './myapprovalyears.png'
         with open(path, 'wb') as f:
             try:
                 f.write(rd1.get('genes_approved'))
             except TypeError:
-                return ("No image in the database\n")
+                return ("No image in the database. Please use a POST route first.\n")
             f.write(rd1.get('genes_approved'))
         return send_file(path, mimetype='image/png', as_attachment=True)
     elif request.method == 'DELETE':
-        try:
-            len(rd1.keys()) >= 1
-        except Exception:
-            return ("No image in the database to delete")
+        if len(rd1.keys()) < 1:
+            return ("No image in the database to delete. Please use a POST route first.\n")
         rd1.flushdb()
-        return f'Image deleted, there are {len(rd1.keys())} keys in the db\n'
+        return f'Image deleted, there are {len(rd1.keys())} keys in the db.\n'
     else:
-        return 'The method you tried does not work\n'
+        return 'The method you tried does not work.\n'
